refactor(APIM-123): separate facility guarantee controller tests
