+++ src/modules/facility-guarantee/facility-guarantee.controller.get-guarantees.test.ts
@@ -3,8 +3,6 @@ import { GetFacilityGuaranteeGenerator } from '@ukef-test/support/generator/get-
 import { RandomValueGenerator } from '@ukef-test/support/generator/random-value-generator';
 import { when } from 'jest-when';
 
-import { CreateFacilityGuaranteeRequestItem } from './dto/create-facility-guarantee-request.dto';
-import { CreateFacilityGuaranteeResponse } from './dto/create-facility-guarantee-response.dto';
 import { FacilityGuaranteeController } from './facility-guarantee.controller';
 import { FacilityGuaranteeService } from './facility-guarantee.service';
 
@@ -20,16 +18,12 @@ describe('FacilityGuaranteeController', () => {
   });
 
   let getFacilityGuaranteesService: jest.Mock;
-  let facilityGuaranteeServiceCreateGuaranteeForFacility: jest.Mock;
   let controller: FacilityGuaranteeController;
 
   beforeEach(() => {
     const facilityGuaranteeService = new FacilityGuaranteeService(null, null, null, null);
     getFacilityGuaranteesService = jest.fn();
     facilityGuaranteeService.getGuaranteesForFacility = getFacilityGuaranteesService;
-
-    facilityGuaranteeServiceCreateGuaranteeForFacility = jest.fn();
-    facilityGuaranteeService.createGuaranteeForFacility = facilityGuaranteeServiceCreateGuaranteeForFacility;
 
     controller = new FacilityGuaranteeController(facilityGuaranteeService);
   });
@@ -53,38 +47,6 @@ describe('FacilityGuaranteeController', () => {
       const guarantees = await controller.getGuaranteesForFacility({ facilityIdentifier });
 
       expect(guarantees).toStrictEqual(guaranteesFromService);
-    });
-  });
-
-  describe('createGuaranteeForFacility', () => {
-    const facilityIdentifier = valueGenerator.facilityId();
-    const limitKey = valueGenerator.acbsPartyId();
-    const guarantorParty = valueGenerator.acbsPartyId();
-    const guaranteeTypeCode = valueGenerator.stringOfNumericCharacters({ maxLength: 3 });
-    const effectiveDate = valueGenerator.dateOnlyString();
-    const guaranteeExpiryDate = valueGenerator.dateOnlyString();
-    const maximumLiability = valueGenerator.nonnegativeFloat();
-
-    const newGuarantee = new CreateFacilityGuaranteeRequestItem(
-      facilityIdentifier,
-      effectiveDate,
-      limitKey,
-      guaranteeExpiryDate,
-      maximumLiability,
-      guarantorParty,
-      guaranteeTypeCode,
-    );
-
-    it('creates a guarantee for the facility with the service from the request body', async () => {
-      await controller.createGuaranteeForFacility({ facilityIdentifier }, [newGuarantee]);
-
-      expect(facilityGuaranteeServiceCreateGuaranteeForFacility).toHaveBeenCalledWith(facilityIdentifier, newGuarantee);
-    });
-
-    it('returns the facility identifier if creating the guarantee succeeds', async () => {
-      const response = await controller.createGuaranteeForFacility({ facilityIdentifier }, [newGuarantee]);
-
-      expect(response).toStrictEqual(new CreateFacilityGuaranteeResponse(facilityIdentifier));
     });
   });
 });
